ti_Tag should always be treated as ULONG
git-svn-id: https://svn.aros.org/svn/aros/trunk@51766 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/arossupport/libfindtagitem.c
@@ -12,7 +12,7 @@ struct TagItem *LibFindTagItem(Tag tagValue, const struct TagItem *tagList)
 
     while ((tag = LibNextTagItem(&tstate)))
     {
-	if (tag->ti_Tag == tagValue)
+	if ((ULONG)tag->ti_Tag == (ULONG)tagValue)
 	    return tag;
     }
